chore: refactor check package
The `check` package is only used from the `runner` package. It makes
sense to use `TestRunContext` and other types in `check` but that is not
possible due to circular imports. Make `check` part of `runner` instead.

This is a change in preparation of further work.
