+++ runner/check_status.go
@@ -1,11 +1,12 @@
 // Copyright 2024 OWASP CRS Project
 // SPDX-License-Identifier: Apache-2.0
 
-package check
+package runner
 
 import (
-	"github.com/rs/zerolog/log"
 	"slices"
+
+	"github.com/rs/zerolog/log"
 )
 
 var negativeExpectedStatuses = []int{200, 404, 405}
