+++ runner/run_test.go
@@ -20,18 +20,12 @@ import (
 	"github.com/rs/zerolog/log"
 	"github.com/stretchr/testify/suite"
 
-	"github.com/coreruleset/go-ftw/check"
 	"github.com/coreruleset/go-ftw/config"
 	"github.com/coreruleset/go-ftw/ftwhttp"
 	"github.com/coreruleset/go-ftw/output"
 	"github.com/coreruleset/go-ftw/test"
 	"github.com/coreruleset/go-ftw/waflog"
 )
-
-var logText = `[Tue Jan 05 02:21:09.637165 2021] [:error] [pid 76:tid 139683434571520] [client 172.23.0.1:58998] [client 172.23.0.1] ModSecurity: Warning. Pattern match "\\\\b(?:keep-alive|close),\\\\s?(?:keep-alive|close)\\\\b" at REQUEST_HEADERS:Connection. [file "/etc/modsecurity.d/owasp-crs/rules/REQUEST-920-PROTOCOL-ENFORCEMENT.conf"] [line "339"] [id "920210"] [msg "Multiple/Conflicting Connection Header Data Found"] [data "close,close"] [severity "WARNING"] [ver "OWASP_CRS/3.3.0"] [tag "application-multi"] [tag "language-multi"] [tag "platform-multi"] [tag "attack-protocol"] [tag "paranoia-level/1"] [tag "OWASP_CRS"] [tag "capec/1000/210/272"] [hostname "localhost"] [uri "/"] [unique_id "X-PNFSe1VwjCgYRI9FsbHgAAAIY"]
-[Tue Jan 05 02:21:09.637731 2021] [:error] [pid 76:tid 139683434571520] [client 172.23.0.1:58998] [client 172.23.0.1] ModSecurity: Warning. Match of "pm AppleWebKit Android" against "REQUEST_HEADERS:User-Agent" required. [file "/etc/modsecurity.d/owasp-crs/rules/REQUEST-920-PROTOCOL-ENFORCEMENT.conf"] [line "1230"] [id "920300"] [msg "Request Missing an Accept Header"] [severity "NOTICE"] [ver "OWASP_CRS/3.3.0"] [tag "application-multi"] [tag "language-multi"] [tag "platform-multi"] [tag "attack-protocol"] [tag "OWASP_CRS"] [tag "capec/1000/210/272"] [tag "PCI/6.5.10"] [tag "paranoia-level/2"] [hostname "localhost"] [uri "/"] [unique_id "X-PNFSe1VwjCgYRI9FsbHgAAAIY"]
-[Tue Jan 05 02:21:09.638572 2021] [:error] [pid 76:tid 139683434571520] [client 172.23.0.1:58998] [client 172.23.0.1] ModSecurity: Warning. Operator GE matched 5 at TX:anomaly_score. [file "/etc/modsecurity.d/owasp-crs/rules/REQUEST-949-BLOCKING-EVALUATION.conf"] [line "91"] [id "949110"] [msg "Inbound Anomaly Score Exceeded (Total Score: 5)"] [severity "CRITICAL"] [ver "OWASP_CRS/3.3.0"] [tag "application-multi"] [tag "language-multi"] [tag "platform-multi"] [tag "attack-generic"] [hostname "localhost"] [uri "/"] [unique_id "X-PNFSe1VwjCgYRI9FsbHgAAAIY"]
-[Tue Jan 05 02:21:09.647668 2021] [:error] [pid 76:tid 139683434571520] [client 172.23.0.1:58998] [client 172.23.0.1] ModSecurity: Warning. Operator GE matched 5 at TX:inbound_anomaly_score. [file "/etc/modsecurity.d/owasp-crs/rules/RESPONSE-980-CORRELATION.conf"] [line "87"] [id "980130"] [msg "Inbound Anomaly Score Exceeded (Total Inbound Score: 5 - SQLI=0,XSS=0,RFI=0,LFI=0,RCE=0,PHPI=0,HTTP=0,SESS=0): individual paranoia level scores: 3, 2, 0, 0"] [ver "OWASP_CRS/3.3.0"] [tag "event-correlation"] [hostname "localhost"] [uri "/"] [unique_id "X-PNFSe1VwjCgYRI9FsbHgAAAIY"]`
 
 var testConfigMap = map[string]string{
 	"BaseConfig": `---
@@ -103,6 +97,7 @@ type runTestSuite struct {
 	ts           *httptest.Server
 	dest         *ftwhttp.Destination
 	tempFileName string
+	context      *TestRunContext
 }
 
 func (s *runTestSuite) SetupSuite() {
@@ -187,6 +182,9 @@ func (s *runTestSuite) TearDownTest() {
 
 func (s *runTestSuite) BeforeTest(_ string, name string) {
 	s.cfg = config.NewDefaultConfig()
+	s.context = &TestRunContext{
+		Config: s.cfg,
+	}
 	// setup test webserver (not a waf)
 	s.newTestServer(logText)
 	if s.logFilePath != "" {
@@ -533,11 +531,11 @@ func (s *runTestSuite) TestIsolatedSanity() {
 			},
 		},
 	}
-	err := RunStage(rc, &check.FTWCheck{}, types.Test{}, stage)
+	err := RunStage(rc, &FTWCheck{}, types.Test{}, stage)
 	s.ErrorContains(err, "'isolated' is only valid if 'expected_ids' has exactly one entry")
 
 	stage.Output.Log.ExpectIds = []uint{1, 2}
-	err = RunStage(rc, &check.FTWCheck{}, types.Test{}, stage)
+	err = RunStage(rc, &FTWCheck{}, types.Test{}, stage)
 	s.ErrorContains(err, "'isolated' is only valid if 'expected_ids' has exactly one entry")
 }
 
@@ -677,7 +675,7 @@ func (s *runTestSuite) TestEncodedRequest() {
 		Output:   s.out,
 	}
 	stage := s.ftwTests[0].Tests[0].Stages[0]
-	_check, err := check.NewCheck(s.cfg)
+	_check, err := NewCheck(s.context)
 	s.T().Cleanup(func() { _ = _check.Close() })
 	s.Require().NoError(err)
 
@@ -701,7 +699,7 @@ func (s *runTestSuite) TestEncodedRequest_InvalidEncoding() {
 		Output:   s.out,
 	}
 	stage := s.ftwTests[0].Tests[0].Stages[0]
-	_check, err := check.NewCheck(s.cfg)
+	_check, err := NewCheck(s.context)
 	s.T().Cleanup(func() { _ = _check.Close() })
 	s.Require().NoError(err)
